Split up SparkeyTest into unit and integration tests.
